// ui/todo/src/actions/types.ts
export interface ADD_TODO_SUCCESS {
    loading: any;
    error: any;
    todos: any;
}
export interface ADD_TODO_FAILURE {
    loading: any;
    error: any;
}
export interface ADD_TODO_STARTED {
    todos: any;
    error: any;
    loading: any;
}
export interface DELETE_TODO {
    loading: any;
    todos: any;
    error: any;
}